Move sidebar layout components into layout/ subdirectory
Organize 13 root-level components that form the sidebar layout
system into resources/js/components/layout/ for clearer
architectural boundaries. Update all import paths.

Components moved: AppContent, AppLogo, AppLogoIcon, AppShell,
AppSidebar, AppSidebarHeader, Breadcrumbs, NavFooter, NavMain,
NavUser, UserInfo, UserMenuContent, PlaceholderPattern

Closes #118
