+++ resources/js/components/layout/AppLogoIcon.vue
